BC-8093 - replace `Logger` with `LegacyLogger` and some clean up
