+++ src/infra/logger/index.ts
@@ -1,7 +1,5 @@
 export * from './interfaces/index.js';
-export * from './legacy-logger.service.js';
 export { LoggerConfig } from './logger.config.js';
 export * from './logger.js';
 export * from './logger.module.js';
-export * from './logging.utils.js';
 export * from './types/index.js';
