+++ src/infra/logger/interfaces/index.ts
@@ -1,2 +1,1 @@
-export * from './legacy-logger.interface.js';
-export * from './loggable.js';
+export * from './logger.interface.js';
