--- src/infra/logger/interfaces/loggable.ts
@@ -1,5 +1,0 @@
-import { ErrorLogMessage, LogMessage, ValidationErrorLogMessage } from '../types/index.js';
-
-export interface Loggable {
-	getLogMessage(): LogMessage | ErrorLogMessage | ValidationErrorLogMessage;
-}
