+++ src/infra/logger/logger.spec.ts
@@ -1,29 +1,27 @@
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
 import { WINSTON_MODULE_PROVIDER } from 'nest-winston';
 import { Logger as WinstonLogger } from 'winston';
-import { Loggable } from './interfaces/index.js';
 import { Logger } from './logger.js';
-import { ErrorLogMessage, LogMessage, ValidationErrorLogMessage } from './types/index.js';
-
-class SampleLoggable implements Loggable {
-	public getLogMessage(): LogMessage | ErrorLogMessage | ValidationErrorLogMessage {
-		const message = {
-			message: 'test message',
-			data: 'test data',
-		};
-
-		return message;
-	}
-}
 
 describe('Logger', () => {
-	let module: TestingModule;
 	let service: Logger;
+	let processStdoutWriteSpy: jest.SpyInstance<
+		boolean,
+		[str: string | Uint8Array, encoding?: BufferEncoding | undefined, cb?: ((err?: Error) => void) | undefined],
+		any
+	>;
+	let processStderrWriteSpy: jest.SpyInstance<
+		boolean,
+		[str: string | Uint8Array, encoding?: BufferEncoding | undefined, cb?: ((err?: Error) => void) | undefined],
+		any
+	>;
 	let winstonLogger: DeepMocked<WinstonLogger>;
 
 	beforeAll(async () => {
-		module = await Test.createTestingModule({
+		const module: TestingModule = await Test.createTestingModule({
 			providers: [
 				Logger,
 				{
@@ -34,91 +32,52 @@ describe('Logger', () => {
 		}).compile();
 
 		service = await module.resolve(Logger);
-		winstonLogger = await module.get(WINSTON_MODULE_PROVIDER);
+		winstonLogger = module.get(WINSTON_MODULE_PROVIDER);
+	});
+
+	beforeEach(() => {
+		processStdoutWriteSpy = jest.spyOn(process.stdout, 'write');
+		processStderrWriteSpy = jest.spyOn(process.stderr, 'write');
 	});
 
 	afterEach(() => {
-		jest.resetAllMocks();
+		processStdoutWriteSpy.mockRestore();
+		processStderrWriteSpy.mockRestore();
 	});
 
-	afterAll(async () => {
-		await module.close();
-	});
+	describe('WHEN error logging', () => {
+		it('should call winstonLogger.error', () => {
+			const error = new Error('custom error');
+			service.error(error.message, error.stack);
+			expect(winstonLogger.error).toBeCalled();
+		});
 
-	describe('info', () => {
-		it('should call info method of WinstonLogger with appropriate message', () => {
-			const loggable = new SampleLoggable();
-			service.setContext('test context');
-
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-			const expectedMessage = expect.objectContaining({
-				message: "{ message: 'test message', data: 'test data' }",
-				context: 'test context',
-			});
-
-			service.info(loggable);
-
-			expect(winstonLogger.info).toBeCalledWith(expectedMessage);
+		it('should work also when providing circular references', () => {
+			const a: { name: string; b?: object | undefined } = { name: 'great' };
+			const b = { a };
+			a.b = b;
+			service.error(a);
+			expect(winstonLogger.error).toBeCalledWith(
+				expect.objectContaining({
+					message: expect.stringMatching(/name.*great/) as string,
+				}),
+			);
 		});
 	});
 
-	describe('warn', () => {
-		it('should call warn method of WinstonLogger with appropriate message', () => {
-			const loggable = new SampleLoggable();
-			service.setContext('test context');
-
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-			const expectedMessage = expect.objectContaining({
-				message: "{ message: 'test message', data: 'test data' }",
-				context: 'test context',
-			});
-
-			service.warning(loggable);
-
-			expect(winstonLogger.warning).toBeCalledWith(expectedMessage);
+	describe('WHEN warn logging', () => {
+		it('should call winstonLogger.warning', () => {
+			const error = new Error('custom error');
+			service.warn(error.message, error.stack);
+			expect(winstonLogger.warning).toBeCalled();
 		});
 	});
 
-	describe('debug', () => {
-		it('should call debug method of WinstonLogger with appropriate message', () => {
-			const loggable = new SampleLoggable();
-			service.setContext('test context');
-
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-			const expectedMessage = expect.objectContaining({
-				message: "{ message: 'test message', data: 'test data' }",
-				context: 'test context',
-			});
-
-			service.debug(loggable);
-
-			expect(winstonLogger.debug).toBeCalledWith(expectedMessage);
-		});
-	});
-
-	describe('setContext', () => {
-		it('should set the context', () => {
-			service.setContext('test');
-
-			// eslint-disable-next-line @typescript-eslint/dot-notation
-			expect(service['context']).toEqual('test');
-		});
-	});
-
-	describe('notice', () => {
-		it('should call notice method of WinstonLogger with appropriate message', () => {
-			const loggable = new SampleLoggable();
-			service.setContext('test context');
-
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-			const expectedMessage = expect.objectContaining({
-				message: "{ message: 'test message', data: 'test data' }",
-				context: 'test context',
-			});
-
-			service.notice(loggable);
-
-			expect(winstonLogger.notice).toBeCalledWith(expectedMessage);
+	describe('WHEN debug logging', () => {
+		it('should call winstonLogger.debug', () => {
+			const error = new Error('custom error');
+			service.debug(error.message, error.stack);
+			expect(winstonLogger.debug).toBeCalled();
 		});
 	});
 });
