+++ src/infra/logger/logger.ts
@@ -1,36 +1,63 @@
 import { Inject, Injectable, Scope } from '@nestjs/common';
 import { WINSTON_MODULE_PROVIDER } from 'nest-winston';
-import { Logger as WinstonLogger } from 'winston';
-import { Loggable } from './interfaces/index.js';
-import { LoggingUtils } from './logging.utils.js';
+import { inspect } from 'util';
+import winston from 'winston';
+import { RequestLoggingBody } from './interfaces/index.js';
+import { ILogger } from './interfaces/logger.interface.js';
 
 @Injectable({ scope: Scope.TRANSIENT })
-export class Logger {
+export class Logger implements ILogger {
+	/**
+	 * This Logger Service can be injected into every Class,
+	 * use setContext() with CustomProviderClass.name that will be added to every log.
+	 * It implements @ILogger which provides the logger methods.
+	 * CAUTION: PREPARE STRINGS AS LOG DATA, DO NOT LOG COMPLEX DATA STRUCTURES
+	 */
 	private context = '';
 
-	public constructor(@Inject(WINSTON_MODULE_PROVIDER) private readonly logger: WinstonLogger) {}
+	public constructor(@Inject(WINSTON_MODULE_PROVIDER) private readonly logger: winston.Logger) {}
 
-	public warning(loggable: Loggable): void {
-		const message = LoggingUtils.createMessageWithContext(loggable, this.context);
-		this.logger.warning(message);
+	public log(message: unknown, context?: string): void {
+		this.logger.info(this.createMessage(message, context));
 	}
 
-	public notice(loggable: Loggable): void {
-		const message = LoggingUtils.createMessageWithContext(loggable, this.context);
-		this.logger.notice(message);
+	public warn(message: unknown, context?: string): void {
+		this.logger.warning(this.createMessage(message, context));
 	}
 
-	public info(loggable: Loggable): void {
-		const message = LoggingUtils.createMessageWithContext(loggable, this.context);
-		this.logger.info(message);
+	public debug(message: unknown, context?: string): void {
+		this.logger.debug(this.createMessage(message, context));
 	}
 
-	public debug(loggable: Loggable): void {
-		const message = LoggingUtils.createMessageWithContext(loggable, this.context);
-		this.logger.debug(message);
+	public http(message: RequestLoggingBody, context?: string): void {
+		this.logger.notice(this.createMessage(message, context));
+	}
+
+	public error(message: unknown, trace?: unknown, context?: string): void {
+		const result = {
+			message,
+			trace,
+		};
+		this.logger.error(this.createMessage(result, context));
 	}
 
 	public setContext(name: string): void {
 		this.context = name;
+	}
+
+	private createMessage(
+		message: unknown,
+		context?: string | undefined,
+	): {
+		message: string;
+		context: string;
+	} {
+		return { message: this.stringifiedMessage(message), context: context ?? this.context };
+	}
+
+	private stringifiedMessage(message: unknown): string {
+		const stringifiedMessage = inspect(message).replace(/\n/g, '').replace(/\\n/g, '');
+
+		return stringifiedMessage;
 	}
 }
